Remove name from custom indexes.
Having the name present causes a problem on the second full data load.
This is because the index name is scoped to the schema, not the table name.

eg. From the first import, we would have in the database:
TABLE artist
INDEX artist_lower_idx

Then during the second import, we would have the above plus additionally:
TABLE artist_tmp

And the "CREATE INDEX IF NOT EXISTS artist_lower_idx" query is skipped,
because the index with that name from first import still exists.

Omitting the name will allow PostgreSQL to generate a unique non-conflicting index name.